requirements.txt
kaggle==1.6.17
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
pyyaml==6.0.2
tqdm==4.66.6
matplotlib==3.9.2

torch==2.5.1
transformers==4.46.3
accelerate==1.1.1